Node controllers: Retry on OperationalError
The sentry tooling helped me see that after we caught an
InvalidRequestError, the next attempt would hit an OperationalError, so
we wouldn't retry after that.

After looking through some logs, I don't see us ever retrying a second
time, so let's also reduced the maximum possible attempts to 5, from 20.

Finally, let's be a little more informative in the log when retries are exhausted.

Signed-off-by: Zack Cerza <zack@redhat.com>